backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
firebase-admin
pyrebase4
python-jose
python-multipart
requests
scikit-learn
pandas
lightgbm
joblib
weasyprint
twilio
beautifulsoup4
selenium
scrapy
numpy
scipy